0.4.0: breaking: allow clients to customize the body serialization logic
adds the concept of a serializer with SerializeBody trait so you can do more than serialize with serde json, such as have no request body, or whatever you want

Request implementation now requires Serializer to be specified, since associated type defaults are unstable. so this is a breaking change mainly due to that. available serializers are Serde and NoBody

it also changes the error type to be generic based on the serialization method, but that shouldn't break anything since it uses serde as the default

also some type signatures have changed in the Client so that might be breaking but i haven't put a ton of thought into that